# Saving memory in the classical calculator by reducing the size of the .rmap