This is relevant for the USA model. Thanks to this optimization it is possible to skip the inefficiencies of the tiling calculator and to get a significant speedup. On the spot machine with 5% of the sites I get
```
# before
| calc_35, maxmem=325.6 GB   | time_sec | memory_mb | counts     |
|----------------------------+----------+-----------+------------|
| total classical            | 641_709  | 1_001     | 261        |
| iter_ruptures              | 280_966  | 0.0       | 862_623    |
| nonplanar contexts         | 167_273  | 0.0       | 225_198    |
| get_poes                   | 92_612   | 0.0       | 28_485_270 |
| planar contexts            | 84_474   | 0.0       | 13_565_889 |
| computing mean_std         | 59_294   | 0.0       | 707_652    |
| building dparam            | 12_419   | 0.0       | 18_512     |
| ClassicalCalculator.run    | 7_184    | 3_541     | 1          |

# after, 4x speedup, 4x less memory
| calc_38, maxmem=86.5 GB    | time_sec | memory_mb | counts     |
|----------------------------+----------+-----------+------------|
| total classical            | 161_056  | 130.0625  | 300        |
| get_poes                   | 86_338   | 0.0       | 13_441_481 |
| computing mean_std         | 46_847   | 0.0       | 274_423    |
| nonplanar contexts         | 8_616    | 0.0       | 76_473     |
| iter_ruptures              | 6_862    | 0.0       | 25_864     |
| total preclassical         | 3_317    | 93.9805   | 306        |
| ClassicalCalculator.run    | 2_758    | 3_497     | 1          |
```